+++ std-bits/table/src/main/java/org/enso/table/data/column/operation/cast/CastProblemBuilder.java
@@ -1,4 +1,4 @@
-package org.enso.table.data.column.operation;
+package org.enso.table.data.column.operation.cast;
 
 public class CastProblemBuilder {
   private int failedConversionsCount = 0;
