+++ std-bits/table/src/main/java/org/enso/table/data/column/storage/MixedStorageFacade.java
@@ -1,7 +1,7 @@
 package org.enso.table.data.column.storage;
 
 import org.enso.table.data.column.builder.object.Builder;
-import org.enso.table.data.column.operation.CastProblemBuilder;
+import org.enso.table.data.column.operation.cast.CastProblemBuilder;
 import org.enso.table.data.column.operation.map.MapOperationProblemBuilder;
 import org.enso.table.data.column.storage.type.AnyObjectType;
 import org.enso.table.data.column.storage.type.StorageType;
